navBar, index, about, album pages
I made the logo bigger in the nav Bar and adjust the spacing so that it
better aligned on the content pages.
I added a tags to the social media links.
I adjusted the layout of the about page. Change the image size.
I change the image size on the album page and styled the position of
the text. I also added in the social media icons.